refactor(reset-password): updated imports
