+++ app/(auth)/reset-password/page.tsx
@@ -1,5 +1,6 @@
 "use client";
 
+import { sendPasswordResetLink } from "@/auth/actions";
 import DialogFormButton from "@/components/shared/buttons/DialogFormButton";
 import CustomFormField, {
   FormFieldType,
@@ -15,7 +16,6 @@ import {
   CardTitle,
 } from "@/components/ui/card";
 import { Form } from "@/components/ui/form";
-import { sendPasswordResetLink } from "@/lib/actions/auth.actions";
 import { formatErrorResponse } from "@/lib/formatters";
 import { zodResolver } from "@hookform/resolvers/zod";
 import { useRouter } from "next/navigation";
